+++ x/gov/client/cli/cli_proposal_test.go
@@ -54,7 +54,7 @@ func (s IntegrationTestSuite) TestCreateProposalUpsertDataRegistry() {
 
 	cmd := cli.GetTxProposalUpsertDataRegistry()
 	clientCtx := val.ClientCtx
-	_, err = clitestutil.ExecTestCLICmd(clientCtx, cmd, []string{
+	out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, []string{
 		fmt.Sprintf("%s", "theKey"),
 		fmt.Sprintf("%s", "theHash"),
 		fmt.Sprintf("%s", "theReference"),
@@ -67,11 +67,12 @@ func (s IntegrationTestSuite) TestCreateProposalUpsertDataRegistry() {
 		fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(100))).String()),
 	})
 	s.Require().NoError(err)
+	fmt.Printf("%s", out.String())
 
 	// Vote Proposal
 	cmd = cli.GetTxVoteProposal()
-	out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, []string{
-		fmt.Sprintf("%d", 2), // Proposal ID
+	out, err = clitestutil.ExecTestCLICmd(clientCtx, cmd, []string{
+		fmt.Sprintf("%d", 1), // Proposal ID
 		fmt.Sprintf("%d", customgovtypes.OptionYes),
 		fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
 		fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
@@ -102,6 +103,37 @@ func (s IntegrationTestSuite) TestCreateProposalSetNetworkProperty() {
 	cmd = cli.GetTxVoteProposal()
 	out, err = clitestutil.ExecTestCLICmd(clientCtx, cmd, []string{
 		fmt.Sprintf("%d", 2), // Proposal ID
+		fmt.Sprintf("%d", customgovtypes.OptionYes),
+		fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
+		fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
+		fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
+		fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(100))).String()),
+	})
+	s.Require().NoError(err)
+	fmt.Printf("%s", out.String())
+}
+
+func (s IntegrationTestSuite) TestCreateProposalCreateRole() {
+	// Query permissions for role Validator
+	val := s.network.Validators[0]
+
+	cmd := cli.GetTxProposalCreateRole()
+	clientCtx := val.ClientCtx
+	out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, []string{
+		fmt.Sprintf("%d", 12345),
+		fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
+		fmt.Sprintf("--%s=%s", cli.FlagWhitelistPerms, "1,2,3"),
+		fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
+		fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
+		fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(s.cfg.BondDenom, sdk.NewInt(100))).String()),
+	})
+	s.Require().NoError(err)
+	fmt.Printf("%s", out.String())
+
+	// Vote Proposal
+	cmd = cli.GetTxVoteProposal()
+	out, err = clitestutil.ExecTestCLICmd(clientCtx, cmd, []string{
+		fmt.Sprintf("%d", 1), // Proposal ID
 		fmt.Sprintf("%d", customgovtypes.OptionYes),
 		fmt.Sprintf("--%s=%s", flags.FlagFrom, val.Address.String()),
 		fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),