+++ x/staking/client/cli/tx.go
@@ -83,7 +83,7 @@ func GetTxClaimValidatorCmd() *cobra.Command {
 // GetTxProposalUnjailValidatorCmd implement cli command for MsgUpsertTokenAlias
 func GetTxProposalUnjailValidatorCmd() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "proposal-unjail-validator",
+		Use:   "proposal-unjail-validator hash reference",
 		Short: "Creates an proposal to unjail validator (the from address is the validator)",
 		Args:  cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) error {